test_production_readiness.py
#!/usr/bin/env python3
"""
Production Readiness Test for Snapchat Automation System
Tests all critical components with REAL services and credentials
"""

import os
import sys
import asyncio
import logging
from datetime import datetime

# Load environment variables from .env file
try:
    from dotenv import load_dotenv
    load_dotenv('.env')
    print("✅ Environment variables loaded from .env file")
except ImportError:
    print("⚠️  python-dotenv not available, using system environment variables")

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class ProductionReadinessTest:
    """Test all critical components for production readiness"""
    
    def __init__(self):
        self.results = {}
        self.overall_status = "UNKNOWN"
        
    def test_environment_variables(self):
        """Test all required environment variables are set"""
        logger.info("🔧 Testing Environment Variables...")
        
        required_vars = [
            'TELEGRAM_BOT_TOKEN',
            'TWILIO_ACCOUNT_SID', 
            'TWILIO_AUTH_TOKEN',
            'DATABASE_URL',
            'REDIS_URL'
        ]
        
        missing_vars = []
        for var in required_vars:
            if not os.getenv(var):
                missing_vars.append(var)
                
        if missing_vars:
            self.results['environment'] = {
                'status': 'FAIL',
                'message': f'Missing environment variables: {missing_vars}'
            }
            return False
        else:
            self.results['environment'] = {
                'status': 'PASS',
                'message': 'All required environment variables present'
            }
            return True
            
    def test_telegram_bot(self):
        """Test Telegram bot connection"""
        logger.info("🤖 Testing Telegram Bot Connection...")
        
        try:
            import telegram
            
            bot_token = os.getenv('TELEGRAM_BOT_TOKEN')
            if not bot_token:
                self.results['telegram'] = {
                    'status': 'FAIL',
                    'message': 'TELEGRAM_BOT_TOKEN not found'
                }
                return False
                
            # Test bot connection
            bot = telegram.Bot(token=bot_token)
            me = asyncio.run(bot.get_me())
            
            self.results['telegram'] = {
                'status': 'PASS',
                'message': f'Bot connected: @{me.username} ({me.first_name})'
            }
            return True
            
        except Exception as e:
            self.results['telegram'] = {
                'status': 'FAIL',
                'message': f'Telegram connection failed: {str(e)}'
            }
            return False
            
    def test_twilio_sms(self):
        """Test Twilio SMS service"""
        logger.info("📱 Testing Twilio SMS Service...")
        
        try:
            from twilio.rest import Client
            
            account_sid = os.getenv('TWILIO_ACCOUNT_SID')
            auth_token = os.getenv('TWILIO_AUTH_TOKEN')
            
            if not account_sid or not auth_token:
                self.results['twilio'] = {
                    'status': 'FAIL',
                    'message': 'Twilio credentials not found'
                }
                return False
                
            # Test Twilio connection
            client = Client(account_sid, auth_token)
            account = client.api.account.fetch()
            
            self.results['twilio'] = {
                'status': 'PASS',
                'message': f'Twilio connected: {account.friendly_name} ({account.status})'
            }
            return True
            
        except Exception as e:
            self.results['twilio'] = {
                'status': 'FAIL',
                'message': f'Twilio connection failed: {str(e)}'
            }
            return False
            
    def test_database_connection(self):
        """Test PostgreSQL database connection"""
        logger.info("🗄️ Testing Database Connection...")
        
        try:
            import psycopg2
            from urllib.parse import urlparse
            
            database_url = os.getenv('DATABASE_URL')
            if not database_url:
                self.results['database'] = {
                    'status': 'FAIL',
                    'message': 'DATABASE_URL not found'
                }
                return False
                
            # Parse and connect to database
            result = urlparse(database_url)
            conn = psycopg2.connect(
                host=result.hostname,
                port=result.port,
                database=result.path[1:],
                user=result.username,
                password=result.password
            )
            
            # Test database operations
            cursor = conn.cursor()
            cursor.execute('SELECT version();')
            version = cursor.fetchone()[0]
            
            cursor.close()
            conn.close()
            
            self.results['database'] = {
                'status': 'PASS',
                'message': f'Database connected: PostgreSQL {version.split()[1]}'
            }
            return True
            
        except Exception as e:
            self.results['database'] = {
                'status': 'FAIL',
                'message': f'Database connection failed: {str(e)}'
            }
            return False
            
    def test_snapchat_automation(self):
        """Test Snapchat automation components"""
        logger.info("👻 Testing Snapchat Automation Components...")
        
        try:
            # Test stealth creator import
            sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'automation'))
            from automation.snapchat.stealth_creator import SnapchatStealthCreator
            
            # Test initialization
            creator = SnapchatStealthCreator()
            
            # Test profile generation
            profile = creator.generate_stealth_profile()
            
            if profile and profile.username and profile.email:
                self.results['snapchat'] = {
                    'status': 'PASS',
                    'message': f'Automation ready: Profile generated ({profile.username})'
                }
                return True
            else:
                self.results['snapchat'] = {
                    'status': 'FAIL',
                    'message': 'Profile generation failed'
                }
                return False
                
        except Exception as e:
            self.results['snapchat'] = {
                'status': 'FAIL',
                'message': f'Snapchat automation failed: {str(e)}'
            }
            return False
            
    def test_fly_io_integration(self):
        """Test Fly.io Android farm integration"""
        logger.info("☁️ Testing Fly.io Android Farm Integration...")
        
        try:
            # Test Android automation components
            sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'automation'))
            
            # Test farm device detection logic
            farm_host = os.getenv("FLY_ANDROID_HOST", "android-device-farm-prod.fly.dev")
            farm_port = os.getenv("FLY_ANDROID_PORT", "5555")
            
            # Simulate farm device ID
            farm_device_id = f"farm_{farm_host}_{farm_port}"
            
            if "fly.dev" in farm_device_id or farm_device_id.startswith("farm_"):
                self.results['fly_io'] = {
                    'status': 'PASS',
                    'message': f'Fly.io integration ready: {farm_host}:{farm_port}'
                }
                return True
            else:
                self.results['fly_io'] = {
                    'status': 'FAIL',
                    'message': 'Fly.io configuration invalid'
                }
                return False
                
        except Exception as e:
            self.results['fly_io'] = {
                'status': 'FAIL',
                'message': f'Fly.io integration failed: {str(e)}'
            }
            return False
            
    def run_all_tests(self):
        """Run all production readiness tests"""
        logger.info("🚀 Starting Production Readiness Tests...")
        logger.info("=" * 60)
        
        tests = [
            self.test_environment_variables,
            self.test_telegram_bot,
            self.test_twilio_sms,
            self.test_database_connection,
            self.test_snapchat_automation,
            self.test_fly_io_integration
        ]
        
        passed = 0
        total = len(tests)
        
        for test in tests:
            try:
                if test():
                    passed += 1
            except Exception as e:
                logger.error(f"Test failed with exception: {e}")
                
        # Generate overall status
        if passed == total:
            self.overall_status = "✅ PRODUCTION READY"
        elif passed >= total * 0.8:
            self.overall_status = "⚠️ MOSTLY READY (minor issues)"
        else:
            self.overall_status = "❌ NOT READY (critical issues)"
            
        # Print results
        self.print_results(passed, total)
        
        return passed == total
        
    def print_results(self, passed, total):
        """Print test results in a formatted report"""
        print("\n" + "=" * 60)
        print("🔍 PRODUCTION READINESS TEST RESULTS")
        print("=" * 60)
        
        for test_name, result in self.results.items():
            status_icon = "✅" if result['status'] == 'PASS' else "❌"
            print(f"{status_icon} {test_name.upper()}: {result['message']}")
            
        print("=" * 60)
        print(f"📊 OVERALL RESULTS: {passed}/{total} tests passed")
        print(f"🎯 STATUS: {self.overall_status}")
        print("=" * 60)
        
        if passed == total:
            print("🚀 READY FOR LIVE SNAPCHAT ACCOUNT CREATION!")
        else:
            print("⚠️  Fix the failed tests before production deployment")
            
        print(f"📅 Test completed: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
        print("=" * 60)

if __name__ == "__main__":
    print("🔥 SNAPCHAT AUTOMATION PRODUCTION READINESS TEST")
    print("Testing all critical components with REAL services...")
    print("=" * 60)
    
    tester = ProductionReadinessTest()
    success = tester.run_all_tests()
    
    if success:
        print("\n🎉 ALL SYSTEMS GO! Ready for production account creation.")
        sys.exit(0)
    else:
        print("\n⚠️  Production readiness test failed. Check the results above.")
        sys.exit(1)