requirements.txt
# Comprehensive Requirements for Tinder Automation System
# Generated to fix import path failures and missing dependencies

# Core Python dependencies
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
requests[socks]==2.31.0
aiohttp==3.9.1
httpx==0.25.2
pydantic==1.10.13
typing-extensions==4.8.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Automation and testing
selenium==4.15.2
appium-python-client==2.11.1
uiautomator2==2.16.23
pure-python-adb==0.3.0.dev0
pytest==7.4.3
pytest-asyncio==0.21.1

# Image processing and computer vision
opencv-python==4.8.1.78
Pillow==10.0.1
pytesseract==0.3.10
easyocr==1.7.0

# Machine learning and AI
scikit-learn==1.3.2
tensorflow==2.13.0
bezier==2023.7.28

# Database drivers
psycopg2-binary==2.9.7
sqlalchemy==1.4.50
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0

# Redis and caching
redis==4.6.0
aioredis==2.0.1

# SMS and communication services
twilio==8.10.0
boto3==1.34.0
messagebird==2.0.0

# Telegram bot framework
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Web framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
hvac==1.1.1

# Payment processing
stripe==7.8.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0
toml==0.10.2
pyyaml==6.0.1

# Utilities and helpers
fake-useragent==1.4.0
Faker==20.1.0
schedule==1.2.0
colorama==0.4.6
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# Date and time handling
python-dateutil==2.8.2
quarterly==3.0.0

# Network and proxy
proxies==0.4.0
socks==1.1.2
urllib3==2.0.7

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Optional enhanced dependencies for specific features
# Uncomment as needed:

# Advanced ML/AI (uncomment if using advanced features)
# torch==2.1.0
# torchvision==0.16.0
# transformers==4.35.2

# Advanced image processing (uncomment if needed)
# scikit-image==0.21.0
# imageio==2.31.6

# Advanced automation (uncomment for enhanced features)
# playwright==1.40.0
# undetected-chromedriver==3.5.4

# Cloud services (uncomment if using cloud features)
# google-cloud-storage==2.10.0
# azure-storage-blob==12.19.0
# aws-sdk-pandas==3.5.0

# Performance optimization (uncomment for production)
# uvloop==0.19.0
# orjson==3.9.10
# cython==3.0.6